Add new json encode/decode funs
OTP 27 now has a built in json module. It returns iolists rather than binarys and
by default decodes objects as maps. This requires using the callback support it includes
in order to support the messages erlbutt uses which are records (actually tuples in erlang).
I may rethink and rework all the message stuff to use maps at some point.

It also requires using callbacks for the encoding in order to workaround 's janky
serialization, getting the extra whitespace and indenting in ther for the signatures.