specify elp as the erlang LSP for the project
upgrade to OTP 28, in oreder to use Zed, elp, and the
Claude AI. Claude did a good job with this refactor. I added a size funtion in
utils that distinguishes between binaries and lists, per the advice of
elp (performance?) and this broke some code that was using size in
guards. This is verboten in erlang. Claude rewrote the if statements as
cases with multiple returns. Pretty good for an AI.